finalizando separação de logica, para o arquivo provider
